blog_website/requirements.txt
Flask==3.0.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
bcrypt==4.0.1
pymongo==4.6.0